bin/install_hawk.sh: Removed all guardian related code
Removed all guardian related code as on almost all machines it will
not be present.

Replaced the pgsql data path with variable, so it can be easily
changed in the future with actual detection of the data dir.

Renamed the file, so it is in the same format as uninstall_hawk.sh.
I decided to keep these with install/uninstall prefixes to separate
them from the rest of the hawk scripts, which are used during the
lifetime of the app.

Signed-off-by: Marian Marinov <mm@yuhu.biz>