hawk.conf: Reorder of the parameters in the config and added a lot of comments
Rordered the parameters so the most important params for the sysadmins are on top.

The DB parameters should be autoconfigured so we move them to the bottom.
Service IDs and Service names should not be touched by the sysadmins, so they are
left last in the configuration file.

Introduced one new parameter sqlite=1, so we can switch to SQLite DB by default
if PgSQL is not found.

Decreased the default value of broot_number from 15 to 10, which should be
sufficient for most cases.

Signed-off-by: Marian Marinov <mm@yuhu.biz>