hawk.pl: Added support for comments, ipset and block_script
To get more flexibility we need to be able to use either block_script,
ipset or iptables for blocking.
This was implemented in the do_block() function as the easiest place,
but in the future this would probably be checked once on daemon start.

We have also added the concept of comments, used for all block methods.

Signed-off-by: Marian Marinov <mm@yuhu.biz>