hawk.pl: Move the logic for the blocking command at the start of Hawk Instead of evaluating the logic every time we are blocking IPs, we are now evaluating this one time by creating an array @block_cmd that holds all parameters for the command with the first parameter, being the index in the @block_cmd which should hold the IP.
As all commands receive the comment as last parameter, the comment is
added directly in the do_block().

Once do_block() is called it will shift the first parameter of the array
and set the IP at the proper place(based on the shifted parameter).

Before we start the main logic for the block_cmd we are verifying that
block_comments is initialized, to reduce the calls to defined() on that
parameter.

In order to use a block_script, the option should be defined and the
file should be executable. Otherwise Hawk will silently rvert to ipset
or iptables.

Signed-off-by: Marian Marinov <mm@yuhu.biz>