bin/hawk-unblock.sh: Significant changes
* Changed the logic, went from if-else to case with functions.
This was done, to make the logic simpler to read and to allow for
further addition of MySQL support.
* sqlite variable is removed as it was replaced by db_type in the
configuration file
* The iptables chain that iptables rules are is now in the config,
so we are now reading it from there.
* Configuration file location is updated, because we moved the file
* pgsql_ublock() now checks if DB name has been extracted from the
config and fails if not present.
* All awk commands within the pgsql_unblock() now redirect stderr
to /dev/null, to prevent unintentional characters in the user, pass
or db name variables.
* db_name is now extracted from the configuration file, instead of
being hardcoded in the script
* Added option for MySQL but not finished.

Signed-off-by: Marian Marinov <mm@yuhu.biz>